Splitted the tmdbcall API wrapper in different files and classes
