+++ tmdbcall/genre.py
@@ -1,0 +1,22 @@
+"""This module holds the Genre class"""
+from .parent import Parent
+
+
+class Genre(Parent):
+    """
+    Class to get Genre Data.
+
+    Attributes:
+        URLS (dict): All relevant urls
+    """
+    URLS = {
+        'genre_list': 'genre/tv/list',
+    }
+
+    def get_genres(self):
+        """Get a list of all tv genres.
+        Builds the URL to the target and makes a request.
+        """
+        target = self.base_uri + self.URLS['genre_list']
+        response = self.make_request(target=target)
+        return response
