+++ tmdbcall/tests/test_tv.py
@@ -1,0 +1,5 @@
+"""This file holds all tests for the TvTMDB class"""
+from django.test import TestCase
+
+import tmdbcall as tmdb
+
